requirement.txt
pip           24.2
psutil        6.1.0
python-dotenv 1.0.1